# Use pool of buffers to reduce GC