Some reference articles:

http://www.akshaydeo.com/blog/2017/12/23/How-did-I-improve-latency-by-700-percent-using-syncPool/
https://blog.questionable.services/article/using-buffer-pools-with-go/
https://blog.cloudflare.com/recycling-memory-buffers-in-go/