@CaptainCodeman 2018-06-18
Thanks for the feedback. How's this?

Benchmark of original:
```
BenchmarkJSON-8    1000000    1821 ns/op    304 B/op    4 allocs/op
```
Bench of PR:
```
BenchmarkJSON-8    1000000    1592 ns/op     96 B/op    2 allocs/op
```
Comparison:
```

benchmark           old ns/op     new ns/op     delta
BenchmarkJSON-8     1821          1592          -12.58%

benchmark           old allocs    new allocs    delta
BenchmarkJSON-8     4             2             -50.00%

benchmark           old bytes     new bytes     delta
BenchmarkJSON-8     304           96            -68.42%
```