@VojtechVitek 2018-06-18
Interesting, so it does escape after all.
```
$ go test -gcflags '-m -l' -bench=. -benchtime 5s

./responder.go:102:25: &buf escapes to heap
./responder.go:102:25: &buf escapes to heap
./responder.go:101:6: moved to heap: buf
```