@VojtechVitek 2018-06-18
The performance gain from this PR is very little. For more real use case:

```go
func init() {
    words := make([]string, 200)
    for i, _ := range words {
        words[i] = "word"
    }

    floats := make([]float64, 200)
    for i, _ := range floats {
        floats[i] = 42.0
    }

    obj = Obj{
        ID:      123,
        Name:    "Some data",
        Words:   []string{"one", "two", "three", "four", "five"},
        Numbers: []float64{1.0, 2.0, 3.0, 4.0, 5.0},
        Words:   words,
        Numbers: floats,
    }
    str = `{"ID":123,"Name":"Some data","Words":["one","two","three","four","five"],"Numbers":[1,2,3,4,5]}
`
)
}
```

```
# original
BenchmarkJSON-4   	  200000	     30176 ns/op	    2592 B/op	       5 allocs/op
PASS
ok  	github.com/go-chi/render	6.377s

# sync.Pool
BenchmarkJSON-4   	  200000	     29945 ns/op	     176 B/op	       3 allocs/op
PASS
ok  	github.com/go-chi/render	6.443s
```