@VojtechVitek 2018-06-18
How about if we don't use the buffer at all?

We could marshal into `w` directly and `panic()` json `.Encode()` error (which is unlikely, since it happens on `v` being a channel/func/complex type only).

```
BenchmarkJSON-4   	  300000	     29069 ns/op	     176 B/op	       3 allocs/op
PASS
ok  	github.com/go-chi/render	9.041s
```

```go
func JSON(w http.ResponseWriter, r *http.Request, v interface{}) {
	enc := json.NewEncoder(w)
	enc.SetEscapeHTML(true)
	if err := enc.Encode(v); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		panic(err)
	}

	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	if status, ok := r.Context().Value(StatusCtxKey).(int); ok {
		w.WriteHeader(status)
	}
}
```