@CaptainCodeman 2018-06-18
> The performance gain from this PR is very little. For more real use case

It's more about reducing allocations to save the _eventual_ GC work which the B/op in your example shows to be less (176 vs 2592). I don't know how to have a benchmark show this offhand - this was all based on the results from running it over time (similar to the articles linked to)