@CaptainCodeman 2018-06-19
Wouldn't we lose the ability to set the http status if it wasn't buffered?