@VojtechVitek 2018-06-19
> It's more about reducing allocations to save the eventual GC work which the B/op in your example shows to be less (176 vs 2592).

I understand. However, GC will collect unused `sync.Pool` items too. So unless you run a REST API application that calls `render.JSON()` in a loop and does nothing else, there won't be much of a performance improvement. The bottle neck is always somewhere else.

> Wouldn't we lose the ability to set the http status if it wasn't buffered?

Good question. Let me try