@VojtechVitek 2018-06-19
I tried rendering struct with `chan int` type using `render.JSON()` to cause the very unlikely error:

The handler panicked, as expected (people should use recoverer middleware anyway, ie. `go-chi/chi/middleware.Recoverer`).
The REST client got HTTP 500 with `handler returned wrong body: got Internal Server Error` body.

That's totally fine in my opinion. Same error code, but less details exposed to the REST customer.

```
$ go test
Panic: json: unsupported type: chan int
goroutine 20 [running]:
runtime/debug.Stack(0x28, 0x0, 0x0)
	/usr/local/go/src/runtime/debug/stack.go:24 +0xa7
runtime/debug.PrintStack()
	/usr/local/go/src/runtime/debug/stack.go:16 +0x22
github.com/go-chi/chi/middleware.Recoverer.func1.1(0xc4201bc100, 0x145e720, 0xc4201ca000)
	/Users/vojtechvitek/go/src/github.com/go-chi/chi/middleware/recoverer.go:28 +0x1b6
panic(0x1395540, 0xc42015e2d0)
	/usr/local/go/src/runtime/panic.go:505 +0x229
github.com/go-chi/render.JSON(0x145e720, 0xc4201ca000, 0xc4201bc100, 0x13da7e0, 0xc42015a230)
	/Users/vojtechvitek/go/src/github.com/go-chi/render/responder.go:105 +0x1c7
github.com/go-chi/render.TestJSON.func1(0x145e720, 0xc4201ca000, 0xc4201bc100)
	/Users/vojtechvitek/go/src/github.com/go-chi/render/responder_test.go:17 +0x6f
net/http.HandlerFunc.ServeHTTP(0x142bcf8, 0x145e720, 0xc4201ca000, 0xc4201bc100)
	/usr/local/go/src/net/http/server.go:1947 +0x44
github.com/go-chi/chi.(*Mux).routeHTTP(0xc420138070, 0x145e720, 0xc4201ca000, 0xc4201bc100)
	/Users/vojtechvitek/go/src/github.com/go-chi/chi/mux.go:415 +0x26d
github.com/go-chi/chi.(*Mux).(github.com/go-chi/chi.routeHTTP)-fm(0x145e720, 0xc4201ca000, 0xc4201bc100)
	/Users/vojtechvitek/go/src/github.com/go-chi/chi/mux.go:351 +0x48
net/http.HandlerFunc.ServeHTTP(0xc42001ee00, 0x145e720, 0xc4201ca000, 0xc4201bc100)
	/usr/local/go/src/net/http/server.go:1947 +0x44
github.com/go-chi/chi/middleware.Recoverer.func1(0x145e720, 0xc4201ca000, 0xc4201bc100)
	/Users/vojtechvitek/go/src/github.com/go-chi/chi/middleware/recoverer.go:35 +0x9f
net/http.HandlerFunc.ServeHTTP(0xc4200f0480, 0x145e720, 0xc4201ca000, 0xc4201bc100)
	/usr/local/go/src/net/http/server.go:1947 +0x44
github.com/go-chi/chi.(*Mux).ServeHTTP(0xc420138070, 0x145e720, 0xc4201ca000, 0xc4201bc000)
	/Users/vojtechvitek/go/src/github.com/go-chi/chi/mux.go:80 +0x1d4
net/http.serverHandler.ServeHTTP(0xc420089790, 0x145e720, 0xc4201ca000, 0xc4201bc000)
	/usr/local/go/src/net/http/server.go:2694 +0xbc
net/http.(*conn).serve(0xc4201a0000, 0x145eb20, 0xc420168140)
	/usr/local/go/src/net/http/server.go:1830 +0x651
created by net/http.(*Server).Serve
	/usr/local/go/src/net/http/server.go:2795 +0x27b
--- FAIL: TestJSON (0.00s)
	responder_test.go:33: handler returned wrong status code: got 500 want 200
	responder_test.go:37: handler returned wrong body: got Internal Server Error
		 want {"ID":123,"Name":"Some data","Words":["one","two","three","four","five"],"Numbers":[1,2,3,4,5]}
FAIL
exit status 1
FAIL	github.com/go-chi/render	0.019s
```