@VojtechVitek 2018-06-19
Let's write to `w` directly and `panic()` on unlikely error. Can you commit on top of your changes? Thanks