@VojtechVitek 2018-06-20
> Let's write to w directly and panic() on unlikely error. Can you commit on top of your changes? Thanks

 @CaptainCodeman pls let me know if you want to finish the PR with the above changes.

Thanks for your contribution. It was nice perf lesson to me :)