@CaptainCodeman 2018-06-20
Will do, just trying an alternative approach with channels so the buffers survive GC. It may be a little too specific to my needs though (using with AppEngine where I can easily limit the # of concurrent requests) so need to play around with it first.